RHEL implementation for kompose role
