+++ molecule/delegated/tests/kompose/main.py
@@ -1,4 +1,4 @@
-from .util.util import get_ansible, get_variable
+from ..util.util import get_ansible, get_variable
 
 testinfra_runner, testinfra_hosts = get_ansible()
 
@@ -7,15 +7,6 @@ def test_kompose_not_installed(host):
     package_name = "kompose"
     package = host.package(package_name)
     assert not package.is_installed
-
-
-def test_kompose_preferences_file(host):
-    f = host.file("/etc/apt/preferences.d/kompose")
-    assert f.exists
-    assert f.user == "root"
-    assert f.group == "root"
-    assert f.mode == 0o644
-    assert "Pin-Priority: -1" in f.content_string
 
 
 def test_kompose_binary_exists(host):
