+++ roles/kompose/tasks/install-Debian.yml
@@ -1,0 +1,19 @@
+---
+- name: Ensure kompose is not installed
+  become: true
+  ansible.builtin.apt:
+    name: kompose
+    state: absent
+    lock_timeout: "{{ apt_lock_timeout | default(300) }}"
+
+- name: Block installation of kompose package
+  become: true
+  ansible.builtin.copy:
+    content: |
+      Package: kompose
+      Pin: version *
+      Pin-Priority: -1
+    mode: 0644
+    dest: /etc/apt/preferences.d/kompose
+    owner: root
+    group: root
