+++ roles/kompose/tasks/install-RedHat.yml
@@ -1,0 +1,20 @@
+---
+- name: Ensure kompose is not installed
+  become: true
+  ansible.builtin.yum:
+    name: kompose
+    state: absent
+
+- name: Block installation of kompose package
+  become: true
+  ansible.builtin.copy:
+    content: |
+      [kompose]
+      name=block kompose package installation
+      baseurl=
+      enabled=1
+      gpgcheck=0
+    mode: 0644
+    dest: /etc/yum.repos.d/kompose.repo
+    owner: root
+    group: root
